Add Some Profiling Prints to Reading WIC Data
The parsing function takes about 13 seconds total to parse the 97 page pdf on my computer. I think that's acceptable, but it would be nice if it were faster. About 12 seconds of time is spent in Page.extract_text(). I don't think we can skip that function. Page.extract_tables() is a alternative that mostly works and only takes about 3 seconds, but it skips over the County information so we'd have to run extract_text() anyway to get the county information. We can't assume the counties are all in order, because we know at least some counties are missing from this pdf. We may be able to use Page.crop() to make a smaller page and just pull the county information from there, but I think that's too much specific optimizing that might not work if the PDFs are formatted differently in the future.