Add WIC CSV Data (#27)
* adding WIC.py script and IWIC.pdf

* Reorganize WIC Files

-Move WIC source data pdf to data_folder since it looks like that's where data sources go.
-Move WIC.py to src folder since that's where the rest of the scripts are and rename it to lower case to match the style of the other scripts.

* Rename For Clarity

-Rename PDF so that it's very clear what data is in it without opening it.
-Rename LA_Total_re to County_Total_re. I don't know what LA Total means, but I think it's easier to understand as the County total.

* Get the wic script runnable from the main script

* Add explanatory comment and fix typo

* Add some comments to explain how the data looks and why we have to merge some columns

* Cover WIC Script In Tests

The script takes about 30 seconds to parse the real pdf, so I'm adding a one page excerpt that we can test with to verify correctness while optimizing.

* Fix parse_wic_pdf() to work with both the test and the main path

* Add Some Profiling Prints to Reading WIC Data

The parsing function takes about 13 seconds total to parse the 97 page pdf on my computer. I think that's acceptable, but it would be nice if it were faster. About 12 seconds of time is spent in Page.extract_text(). I don't think we can skip that function. Page.extract_tables() is a alternative that mostly works and only takes about 3 seconds, but it skips over the County information so we'd have to run extract_text() anyway to get the county information. We can't assume the counties are all in order, because we know at least some counties are missing from this pdf. We may be able to use Page.crop() to make a smaller page and just pull the county information from there, but I think that's too much specific optimizing that might not work if the PDFs are formatted differently in the future.

* Clean Up wic.py

-Change the column names to exactly match the column names from the pdf
-Rephrase some comments to be clearer
-Remove some redundant comments

* Update src/wic.py

Update comment with correct name for the PDF

* Fix all the flake8 errors in the new WIC files

* Change variable names to lower case to match style guide

* Refactor how we're using the dataframe so that we don't have to append to the dataframe and overwrite the variable.

* Make the WIC script default to not rerunning if the output is up to date

* Add verbose option to main to turn off some logging by default. We can pass this in by command line parameters later if we want.

* Fix Remaining Flake8 Errors

Co-authored-by: jrinks <carpediem80@gmail.com>