Make WIC Participation Data Available to Frontend (#44)
* Start changing the wic participation data into using multiple dataframes

* Clean up some code that isn't needed

* -Write four CSV files instead of one confusing one.
-Pass the participation data back in so it can be merged in with the census data

* Cover the new WIC code in tests

* Fix flake8 errors

* Fix missing returns

* Set Up Test for Merging WIC Participation Data into Census Data

Changed the WIC participation dataframes to actually have the fips column as the index. This changed the json format of a lot of the fixtures.

Add new test to test merging. This test currently fails because the code isn't implemented yet

* Get merging tests passing

* Change the data type of the wic participation numbers to ints

* Get the main path working

* Fix flake8 and mypy errors

I'm silencing a mypy error on appending lists of different types:
error: Unsupported operand types for + ("List[str]" and "List[int]")
It looks like the fixes to correct the overzealous type error wouldn't be worth implementing: https://github.com/python/mypy/issues/720

* add main test (#45)

* added test_run

* set fail fast to false, allows determination of os/version specific errors

* replaced .loc[] with reindex(), due to key error

* Fix Error in Test test_run()

ValueError: invalid literal for int() with base 10: '1,044'

* Fix flake8 error

Co-authored-by: michaelpkuhn <mickuhn95@gmail.com>