Fix Tests
test_get_food_insecurity_data() was failing with:

ValueError: Number of class have to be an integer greater than 2 and smaller than the number of values to use

This happened because we recently bumped the number of bins from 4 to 5 so we needed more test data to create bins from. We could also solve this by passing in the number of bins from the test, but I'd rather have the tests more closely match the code paths of the main script.