Fix Tests (#67)
* Fix Tests

test_get_food_insecurity_data() was failing with:

ValueError: Number of class have to be an integer greater than 2 and smaller than the number of values to use

This happened because we recently bumped the number of bins from 4 to 5 so we needed more test data to create bins from. We could also solve this by passing in the number of bins from the test, but I'd rather have the tests more closely match the code paths of the main script.

* Remove file_to_json Tests

test_file_to_json() was failing with the expected output not matching the actual output. This started happening because I added rows to the input files so they could be binned with 5 bins.

-Remove the file_to_json test because test_get_food_insecurity_data runs file_to_json over the exact same files. Removing the test prevents further maintenance in the future.
-Remove code in file_to_json module that was only used by the tests. We don't use this code any more because we're merging the data in memory using Wrappers.

* Remove Code to Run file_to_json as a Script

I don't think anyone was doing this